requirements.txt
openai
python-dotenv
asyncio
requests
bs4
fastapi
python-multipart
uvicorn
bcrypt
circle-developer-controlled-wallets
circle-smart-contract-platform
circle-user-controlled-wallets